requirements.txt
asgiref==3.8.1
bcrypt==4.1.3
blinker==1.8.1
click==8.1.7
Deprecated==1.2.14
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Limiter==3.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
limits==3.11.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.0
psycopg2==2.9.9
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.1
rich==13.7.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
wrapt==1.16.0
